fix: unenrolled default intentions not realizable
The default enrollment ``learner_status`` view now considers intended courses from which the requested user has unenrolled as no-longer realizable.
ENT-9839
