+++ enterprise/__init__.py
@@ -2,4 +2,4 @@
 Your project description goes here.
 """
 
-__version__ = "5.4.0"
+__version__ = "5.4.1"
